chore: Update v1.2-beta (#4)
* style: format code

* feat: move session list

Move sessions to a separate page.
Show the other's name in session detail (fixes #6).
Add export report and remove past sessions button (fixes #5).

* fix: modify session detail

Change `name` field (which might be empty) to `username` field.
Remove `edit` and `cancel` buttons for past sessions.
Remove the placeholder for invalid requests.

* fix: format and prune

Fix errors caused by improper formatting.
Refactor, format, type hints, and prune code.
Separate development and production dependencies and configuration.
Set up VSCode workspace and linter/formatter configurations.

* feat: fixes #2


* Merge branch 'main' into session-list

* chore: Fixup of #12 (#13)

* fix: change profile image handling

Change profile image handling process.
Do not limit file size but will resize to 250x250 pixels or less.
Special profile images (admin, no image) will also be saved.
Overwritten images will be deleted.
Let users set profile images during signup.

* fix: format and prune

Remove unnecessary comments and uncomment necessary code.
Refactor settings and code.

---------

+++ backend/users/signals.py
@@ -13,13 +13,18 @@ from users.models import Profile
 if TYPE_CHECKING:
     from typing import Type
 
-    from django.contrib.auth.models import User as _User
+    from django.contrib.auth.models import AbstractUser
 
-User: Type[_User] = cast("Type[_User]", auth.get_user_model())
+User: Type[AbstractUser] = cast("Type[AbstractUser]", auth.get_user_model())
 
 
 @receiver(post_save, sender=User)
-def save_profile(sender: Type[_User], instance: _User, created: bool, **kwargs):
+def save_profile(
+    sender: Type[AbstractUser],
+    instance: AbstractUser,
+    created: bool,
+    **kwargs,
+):  # pylint: disable=W0613
     if created:
         Profile.objects.get_or_create(user=instance)
     profile: Profile = instance.profile  # type: ignore
@@ -27,7 +32,12 @@ def save_profile(sender: Type[_User], instance: _User, created: bool, **kwargs):
 
 
 @receiver(post_save, sender=User)
-def set_permission(sender: Type[_User], instance: _User, created: bool, **kwargs):
+def set_permission(
+    sender: Type[AbstractUser],
+    instance: AbstractUser,
+    created: bool,
+    **kwargs,
+):  # pylint: disable=W0613
     if created:
         assign_perm("scheduler.add_session", instance)
         teacher_group, created = Group.objects.get_or_create(name="teacher")
@@ -35,6 +45,11 @@ def set_permission(sender: Type[_User], instance: _User, created: bool, **kwargs
 
 
 @receiver(post_delete, sender=User)
-def delete_profile(sender: Type[_User], instance: _User, **kwargs):
-    profile: Profile = instance.profile  # type: ignore
-    profile.delete()
+def delete_profile(
+    sender: Type[AbstractUser],
+    instance: AbstractUser,
+    **kwargs,
+):  # pylint: disable=W0613
+    if hasattr(instance, "profile"):
+        profile: Profile = instance.profile
+        profile.delete()
